Demonstrate handling missing values in R.
This single script `handle_missing_values.R` is meant to be read for its
comments and the learner should follow along, typing in the few commands
used into an interactive session with R in a tool like RStudio.
Ultimately, the tutorial comes from a great group of people at UCLA, so
the learner should open up [the original tutorial][uclaTut], if
possible.

[uclaTut]: http://www.ats.ucla.edu/stat/r/faq/missing.htm