Learn to convert blanks to NA in a factor in a dataframe.
When R encounters empty strings "" in a character vector for a
dataframe, it does not automatically convert them to NA.

The character vector is first coerced into a Factor, which has the empty
string as a level.

One can assign to that level the value NA. This gets rid of the level
entirely, and Factor vector now has NA in the places where there used to
be an empty string.

This can be helpful when writing transformations/mutations that need to
handle NA.